requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
cloudinary==1.40.0
crispy-bootstrap4==2024.1
Django==5.0.6
django-allauth==0.63.3
django-autoslug==1.9.9
django-cloudinary-storage==0.3.0
django-countries==7.6.1
django-crispy-forms==2.2
django-extensions==3.2.3
django-js-asset==2.2.0
django-resized==1.0.2
django-richtextfield==1.6.2
django-summernote==0.8.20.0
pathspec==0.12.1
pillow==10.3.0
pytz==2024.1
sqlparse==0.5.0
stripe==10.1.0
whitenoise==5.3.0
